test: split affected graph classification coverage
